Fusion of code of cmssw for JetCorrection uncertainties in the header file. Changed accordingly other code. To be tested.
